feat: add isLoading flag to ad package to manage placeholder
