--- packages/ad/__tests__/android/android.test.js
@@ -1,5 +1,0 @@
-import shared from "../share.native";
-
-describe("On android", () => {
-  shared();
-});
