+++ packages/ad/__tests__/android/dom-context.android.test.js
@@ -1,0 +1,5 @@
+import shared from "../dom-context.native";
+
+describe("android", () => {
+  shared();
+});
