+++ packages/ad/__tests__/ios/ad.ios.test.js
@@ -1,0 +1,12 @@
+import shared from "../ad.shared";
+
+jest.mock(
+  "../../src/utils/webview-event-callback-setup",
+  () => "mockErrorHandler"
+);
+jest.mock("../../src/ad-init", () => () => "mockAdInit");
+jest.mock("WebView", () => "WebView");
+
+describe("ios", () => {
+  shared();
+});
