--- packages/ad/__tests__/ios/ios.test.js
@@ -1,5 +1,0 @@
-import shared from "../share.native";
-
-describe("On iOS", () => {
-  shared();
-});
