+++ packages/ad/__tests__/web/ad.web.test.js
@@ -1,0 +1,17 @@
+import {
+  addSerializers,
+  compose,
+  minimalWebTransform,
+  rnwTransform,
+  stylePrinter
+} from "@times-components/jest-serializer";
+import shared from "../ad.shared";
+
+describe("web", () => {
+  addSerializers(
+    expect,
+    compose(stylePrinter, minimalWebTransform, rnwTransform())
+  );
+
+  shared();
+});
